requirements.txt
Flask>=2.3
Flask-Cors>=4.0
Flask-SQLAlchemy>=3.1
PyJWT>=2.8
bcrypt>=4.1
gunicorn>=21.2
requests
